Change bot balance from integer to numeric

+++ migrations/20150813204749-create-bots.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       balance: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC
       },
       character: {
         type: Sequelize.STRING
